Move test as well
